Numerous Changes and Additions
VSYNC is now enabled.

Textures use trilinear filtering and have mipmaps generated.

Shaders have been implemented. (Currently using the default raylib vertex and fragment shader.)

Camera zoom limit is now based on the size of the Earth model. (Panning is an issue due to how the zoom limit is calculated, so I will need to account for panning later.)

Earth model contains the cloud layer now. I also included a material file for it so raylib doesn't cry about it. The material file can also be used for material tweaks later.

FPS is uncapped now, but VSYNC keeps the frame rate close to that of your monitor. Without VSYNC, you can get an idea of the absurd performance (~10,000 FPS on my RTX 4080).

The cloud layer rotation now uses fmodf to ensure it doesn't accumulate into a large value. It is also frame rate independent so the uncapped FPS does not affect it.